+++ src/pgta_composer/FileUtils.h
@@ -2,11 +2,12 @@
 #pragma once
 
 #include <string>
+#include <stdint.h>
 #include <stddef.h>
 
 namespace FileUtils
 {
      bool ReadBinaryFileToString(const std::string &filename, std::string &out);
-     bool WriteBinaryToFile(const std::string &filename, const uint8_t *buffer, off_t size);
+     bool WriteBinaryToFile(const std::string &filename, const uint8_t *buffer, size_t size);
      bool WriteAsciiToFile(const std::string &filename, const std::string &buffer);
 }
